example/requirements.txt
PyOpenGL
git+git://github.com/jstasiak/pysfml-cython.git@11a529e29af8e785ff347eb89861a4ef47454c86
